requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
exceptiongroup==1.3.1
fastapi==0.124.4
fonttools==4.61.1
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.46.0
matplotlib==3.10.8
nltk==3.9.2
numba==0.63.1
numpy==2.2.6
nvidia-nccl-cu12==2.28.9
packagename==0.0.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
scam_job_detector==0.0.1
scikit-learn==1.7.2
scipy==1.15.3
shap==0.49.1
six==1.17.0
slicer==0.0.8
starlette==0.50.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.38.0
wordcloud==1.9.4
xgboost-cpu==3.1.2
